Move components to own files
